requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
faster-whisper==0.9.0
gTTS==2.3.2
requests==2.31.0
pydub==0.25.1
groq==0.4.0
python-dotenv==1.0.0
